--- a/python/tvm/relay/__init__.py
+++ b/python/tvm/relay/__init__.py
@@ -28,7 +28,7 @@ from . import module
 from . import adt
 from . import analysis
 from . import transform
-from .build_module import build, create_executor
+from .build_module import build, create_executor, optimize
 from .transform import build_config
 from . import prelude
 from . import parser