In autotvm, users can extract tuning tasks from models that are imported from relay.testing or other frameworks like tensorflow, however, current task extraction execution path contains unnecessary building of relay module into binary code. This PR is to optimize such an execution path and to save task extraction time.

It essentially exposes another interface from inside RelayBuildModule class, "optimize", that generates an optimized relay module (optimizing by FuseOps pass is essential). This output module can be fed to GraphRuntimeCodegen that goes through relay AST and detects corresponding tuning tasks while doing so. The new execution path eliminates the unnecessary tvm::build call, which takes a lot of time, without affecting the original task extraction goal.

Experiments of extracting tuning tasks from relay.testing models:

model | before(s) | after(s) | speedup
--- | --- | --- | ---
resnet-18 | 137.04 | 15.61 | 8.78x
resnet-34 | 137.11 | 15.74 | 8.71x
resnet-50 | 250.77 | 21.88 | 11.46x
resnet-101 | 251.06 | 22.33 | 11.24x
vgg-16 | 53.64 | 0.58 | 92.48x
vgg-19 | 53.63 | 0.58 | 92.46x
mobilenet | 108.7 | 18.42 | 5.9x
squeezenet_v1.1 | 89.62 | 0.53 | 169.09x
inception_v3 | 325.19 | 34.51 | 9.42x

The extracted tasks have been verified with original results.