@kevinthesun 2019-10-21
A bit curious about how topi calls are collected during optimization，is it through certain pass?