@liaha 2019-10-21
> A bit curious about how topi calls are collected during optimization，is it through certain pass?

AFAIK, the topi calls are collected during graphcodegen, however, the graphcodegen can only accept some optimized relay module from BuildModule.Optimize function, otherwise graphcodegen will fail.