@tqchen 2019-10-24
@ZihengJiang  @merrymercy can you also take a look into this PR?